feat(visual-editing): export `createDataAttribute`
